requirements.txt
playwright==1.40.0
pytest==7.4.0
pytest-playwright==0.4.4
python-dotenv
pytest-html==3.2.0
pytest-xdist==3.3.1
pyperclip==1.11.0